% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBetaAr1Ns.R
\docType{data}
\name{dataBetaAr1Ns}
\alias{dataBetaAr1Ns}
\title{Sample Data under Beta Distribution for Small Area Estimation using Hierarchical Bayesian Method for Rao Yu Model with Non Sampled Area}
\format{
A data frame with 100 row and 6 column:
\describe{
  \item{ydi}{Direct Estimation of y}
  \item{area}{Area (domain) of the data}
  \item{period}{Period (subdomain) of the data}
  \item{vardir}{Sampling Variance of y}
  \item{xdi1}{Auxiliary variable of xdi1}
  \item{xdi2}{Auxiliary variable of xdi2}
}
}
\usage{
dataBetaAr1Ns
}
\description{
\enumerate{
  \item A dataset under Beta Distribution to simulate Small Area Estimation using Hierarchical Bayesian method for Rao-Yu Model with Non-sampled Area
  \item This data contains NA values that indicates no sampled in at least one area.
 }
}
\keyword{datasets}
